Started more UI design
I’m still not entirely convinced a switch to Swift is in our best
interests, as that would make cross platform support much more
difficult than I’d like it to be, and Xcode is a slow and insidious
killer.  I like the UI capabilities of Swift much more than Java, but
Java makes a lot more sense functionally.